Move poc code into dedicate location
